final version which have passed with 125% score including the bonus part <3.

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: oatiya <oatiya@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 14:04:31 by oatiya            #+#    #+#             */
-/*   Updated: 2025/10/20 14:04:32 by oatiya           ###   ########.fr       */
+/*   Updated: 2025/10/28 12:10:11 by oatiya           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
